dlthub/pyproject.toml
[project]
name = "dlthub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.2"
dependencies = [
    "dlt[filesystem]>=1.20.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
